Fix native name for Finnish language
The current version was a direct translation of the English verb
"finish" in infinitive form. This new version is the real name of
the Finnish language in Finnish.

Also removed the abbreviation "FI" because for example "Deutsch"
(German) did not have it. It's not clear if it was meant as an
abbreviation for the language, for the country or both, but in the case
of Finnish, it's not needed because Finnish is basically only spoken in
Finland.